refactor tests to match new cell structure
