+++ android/sdl_android/src/androidTest/java/com/smartdevicelink/test/Test.java
@@ -4,10 +4,10 @@ import android.graphics.Color;
 import android.util.Log;
 
 import com.smartdevicelink.R;
-import com.smartdevicelink.SdlConnection.SdlSession2;
 import com.smartdevicelink.managers.file.filetypes.SdlArtwork;
 import com.smartdevicelink.managers.lockscreen.LockScreenConfig;
-import com.smartdevicelink.managers.screen.menu.cells.MenuCell;
+import com.smartdevicelink.managers.screen.menu.MenuCell;
+import com.smartdevicelink.managers.screen.menu.MenuSelectionListener;
 import com.smartdevicelink.protocol.SdlProtocol;
 import com.smartdevicelink.protocol.enums.FunctionID;
 import com.smartdevicelink.proxy.SdlProxyBase;
@@ -396,8 +396,6 @@ public class Test {
 	public static final LightControlData               GENERAL_LIGHTCONTROLDATA               = new LightControlData();
 	public static final HMISettingsControlData         GENERAL_HMISETTINGSCONTROLDATA         = new HMISettingsControlData();
 	public static final SdlArtwork                     GENERAL_ARTWORK                        = new SdlArtwork();
-	public static final MenuCell                       GENERAL_MENUCELL                       = new MenuCell(GENERAL_STRING);
-	public static final int                            GENERAL_MENU_MAX_ID                    = 2000000000;
 
 	public static final HMICapabilities                GENERAL_HMICAPABILITIES                = new HMICapabilities();
 
@@ -454,6 +452,13 @@ public class Test {
 	public static final List<WeatherData>               GENERAL_WEATHERDATA_LIST               = Arrays.asList(GENERAL_WEATHERDATA);
 	public static final List<WeatherAlert>              GENERAL_WEATHERALERT_LIST              = Arrays.asList(GENERAL_WEATHERALERT);
 	public static final List<NavigationInstruction>     GENERAL_NAVIGATION_INSTRUCTION_LIST    = Arrays.asList(GENERAL_NAVIGATION_INSTRUCTION);
+	public static final int                             GENERAL_MENU_MAX_ID                    = 2000000000;
+	public static final MenuCell                        GENERAL_MENUCELL                       = new MenuCell(GENERAL_STRING, new MenuSelectionListener() {
+		@Override
+		public void onTriggered(TriggerSource trigger) {
+			//
+		}
+	});
 	public static final List<MenuCell>                  GENERAL_MENUCELL_LIST                  = Arrays.asList(GENERAL_MENUCELL);
